+++ extra/CHANGELOG.md
@@ -3,3 +3,8 @@
 * version 1.0.0 June 2019
 	* First release.
 
+* Version 2.0.0 November 2019
+	* Added Model 2 for Fourteen segment displays
+	* Added Sixteen segment display support
+	* added Seven Segment display support
+